Display change so Fetchable creatures is now an int

+++ src/PTUI.java
@@ -22,7 +22,7 @@ public class PTUI {
         ArrayList<Double> stats = Statistics.compute(oneDrops, twoDrops, threeDrops);
 
         System.out.println();
-        System.out.println("Number of Fetchable Creatures: " + stats.get(0));
+        System.out.println("Number of Fetchable Creatures: " + stats.get(0).intValue());
         System.out.println("Average CMC for Creatures: " + stats.get(1));
         System.out.println("P[0] Creatures: " + stats.get(2));
         System.out.println("P[1] Creatures: " + stats.get(3));
